emu.gb.Core: Fix UnknownOpcodeException msg - explicit cast from promoted int to fix address in message

+++ src/net/coderobe/retroswing/emu/gb/Core.java
@@ -9,7 +9,7 @@ public class Core {
 		try {
 			opcodes.get(mmu.ram.get(mmu.PC++)).exec();
 		} catch(NullPointerException e) {
-			throw new UnknownOpcodeException("Unknown opcode "+String.format("0x%02X", mmu.ram.get((short)(mmu.PC-1)))+" at address "+String.format("0x%04X", mmu.PC-1));
+			throw new UnknownOpcodeException("Unknown opcode "+String.format("0x%02X", mmu.ram.get((short)(mmu.PC-1)))+" at address "+String.format("0x%04X", (short)(mmu.PC-1)));
 		}
 	}
 	private final Map<Byte, Opcode> opcodes = new HashMap<Byte, Opcode>(){{
